don't override default support files when running init

+++ lib/integra/commands.rb
@@ -37,8 +37,10 @@ module Integra
 
       support_dest = './features/support/'
       Dir.glob(Integra.gem_libdir + "/template/features/support/*.rb").each do |source|
-        FileUtils.cp(source, File.expand_path(support_dest))
-        Integra.logger.action('create', "#{support_dest}#{File.basename(source)}")
+        unless File.exists?("#{File.expand_path(support_dest)}#{source}")
+          FileUtils.cp(source, File.expand_path(support_dest))
+          Integra.logger.action('create', "#{support_dest}#{File.basename(source)}")
+        end
       end
     end
